requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.0
spotipy==2.23.0
yandex-music==2.2.0
pydantic==2.5.0
redis==5.0.1
pytz==2023.3.post1
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
